Updated layout.html and index.html.
Added a Nav Bar Block to Layout as it wont be shown on login or signup screen but moved navbar code to Index. Also included a div class for both sections, workouts and routines with proper sizing. The same card used to create the sections could be used for the workout listing with the change of the colour to #212121 to give the elevation effect.